obstacle avoidance added

+++ src/functions/onAgentCallback.js
@@ -1,6 +1,9 @@
 import {roundedPosition} from "../classes/Position.js";
 import AgentBelief from "../classes/AgentBelief.js";
 
+const AGENT_PROBABILITY_DECAY = 0.1;
+const AGENT_PROBABILITY_THRESHOLD = 0.7;
+
 /**
  * @param {AgentData[]} agents
  * @param {BeliefSet} beliefs
@@ -23,10 +26,14 @@ export async function onAgentCallback_simple(agents,beliefs) {
 			// infer agent belief
 
 			// should consider the believed intention and execute a step of it
-			// in this implementation we will just forget about agents not sensed
-
-			beliefs.deleteAgentBelief(existing_belief);
-
+			// in this implementation we will just reduce the probability of an agent being there
+			
+			existing_belief.probability = existing_belief.probability - AGENT_PROBABILITY_DECAY;
+			existing_belief.time = Date.now();
+			
+			if(existing_belief.probability <= AGENT_PROBABILITY_THRESHOLD){
+				beliefs.deleteAgentBelief(existing_belief);
+			}
 		}
 	}
